requirements.txt
requests>=2.31.0
sqlite3-api>=2.0.0
tqdm>=4.65.0
schedule>=1.2.0
python-dateutil>=2.8.2
